+++ NEWS.md
@@ -3,6 +3,13 @@
   * added grid_start argument option to sar_average and sar_multi
   * edited grid_start to ensure very small starting par values are always included
   * edited how grid_start works and have added a grid_n argument
+  * changed the negative exponential model fitting process so that the z parameter can
+    be any number rather than constrained between 0 and 1.
+  * changed the asymptotic model fitting process so that a negative z-value cannot be returned
+  * bug corrections in the confidence interval function, and adapting it to work with
+    grid_start
+  * Changing model plotting to plot smooth curves by creating 1000 fitted values
+    using fitted parameters
 
 ## Version 1.2.2
   * bug fix - no AICc option in confidence interval function